Changed isNotClosing to isOpen because if isOpen is false inside of any of the using methods, it throws an exception.
