+++ API/src/main/java/io/github/jwdeveloper/tiktok/data/events/TikTokLinkMicArmiesEvent.java
@@ -27,6 +27,7 @@ import io.github.jwdeveloper.tiktok.annotations.EventType;
 import io.github.jwdeveloper.tiktok.data.events.common.TikTokHeaderEvent;
 import io.github.jwdeveloper.tiktok.data.models.LinkMicArmy;
 import io.github.jwdeveloper.tiktok.data.models.Picture;
+import io.github.jwdeveloper.tiktok.messages.enums.LinkMicBattleStatus;
 import io.github.jwdeveloper.tiktok.messages.webcast.WebcastLinkMicArmies;
 import lombok.Getter;
 
@@ -39,8 +40,10 @@ import java.util.List;
 @EventMeta(eventType = EventType.Message)
 public class TikTokLinkMicArmiesEvent extends TikTokHeaderEvent {
     private final Long battleId;
-
-    private final Integer battleStatus;
+    /**
+     true if battle is finished otherwise false
+     */
+    private final boolean finished;
 
     private final Picture picture;
 
@@ -51,14 +54,6 @@ public class TikTokLinkMicArmiesEvent extends TikTokHeaderEvent {
         battleId = msg.getId();
         armies = msg.getBattleItemsList().stream().map(LinkMicArmy::new).toList();
         picture = Picture.map(msg.getImage());
-        battleStatus = msg.getBattleStatus();
-    }
-
-    /**
-     battleStatus of 1 is Ongoing battle & 2 is Finished Battle
-     @return true if battle is finished otherwise false
-     */
-    public boolean isFinished() {
-        return battleStatus == 2;
+        finished = msg.getBattleStatus() == LinkMicBattleStatus.ARMY_FINISHED;
     }
 }
